+++ R/app_03_record_utils.R
@@ -100,8 +100,8 @@ check_for_open_milestones <- function(owner, repo, milestone_names) {
         get_milestone_from_name(owner, repo, milestone_name)
       },
       error = function(e) {
-        debug(.le$logger, glue::glue("Error retrieving Milestones: {e$message}"))
-        rlang::abort(e$message)
+        debug(.le$logger, glue::glue("Error retrieving Milestones: {conditionMessage(e)}"))
+        rlang::abort(conditionMessage(e))
       }
     )
   })
@@ -122,8 +122,8 @@ check_for_open_issues <- function(owner, repo, milestone_names) {
         get_all_issues_in_milestone(owner, repo, milestone_name)
       },
       error = function(e) {
-        debug(.le$logger, glue::glue("Error retrieving Issues from Milestone: {milestone_name}: {e$message}"))
-        rlang::abort(e$message)
+        debug(.le$logger, glue::glue("Error retrieving Issues from Milestone: {milestone_name}: {conditionMessage(e)}"))
+        rlang::abort(conditionMessage(e))
       }
     )
 
@@ -147,8 +147,8 @@ check_for_open_checklists <- function(owner, repo, milestone_names) {
         get_all_issues_in_milestone(owner, repo, milestone_name)
       },
       error = function(e) {
-        debug(.le$logger, glue::glue("Error retrieving Issues from {milestone_name}: {e$message}"))
-        rlang::abort(e$message)
+        debug(.le$logger, glue::glue("Error retrieving Issues from {milestone_name}: {conditionMessage(e)}"))
+        rlang::abort(conditionMessage(e))
       }
     )